+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-Nothing yet :)
+* Allow (bare) trigger attributes to be inline, rather than always on their own line
 
 # v0.2.1